requirements.txt
click==8.0.1
joblib==1.0.1
nltk==3.6.2
PyStemmer==2.0.1
regex==2021.8.21
tqdm==4.62.1
